wired up auth, made schema changes and implemented tokens

+++ internal/store/workout_store_test.go
@@ -147,7 +147,7 @@ func TestCreateWorkout(t *testing.T) {
 	}
 }
 
-func IntPtr(i int) *int {
+func IntPtr(i int64) *int64 {
 	return &i
 }
